requirements.txt
google-generativeai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-adk
faiss-cpu
sentence-transformers
rank_bm25
regex
pillow
python-docx
pypdf
beautifulsoup4
python-dotenv
pandas
openpyxl
google-cloud-aiplatform